+++ examples/preprocessing/plot_ica_from_raw.py
@@ -4,7 +4,7 @@ Compute ICA components on raw data
 ==================================
 
 ICA is fit to MEG raw data.
-The sources matching the EOG are automatically found and displayed.
+The sources matching the eog are automatically found and displayed.
 Subsequently, artifact detection and rejection quality are assessed.
 """
 print(__doc__)
@@ -71,7 +71,7 @@ ica.plot_sources(eog_evoked)  # latent EOG sources + selction
 ica.plot_overlay(eog_evoked)  # overlay raw and clean EOG artifacts
 
 # check the amplitudes do not change
-ica.plot_overlay(raw)  # ECG artifacts remain
+ica.plot_overlay(raw)  # EOG artifacts remain
 
 ###############################################################################
 # To save an ICA solution you can say: