+++ mne/preprocessing/tests/test_ica.py
@@ -373,10 +373,18 @@ def test_ica_additional():
         ica.detect_artifacts(raw, start_find=0, stop_find=50, ecg_ch=ch_name,
                              eog_ch=ch_name, skew_criterion=idx,
                              var_criterion=idx, kurt_criterion=idx)
-    ## score funcs epochs ##
-    idx, scoress = ica.find_bads_ecg(raw)
+
+    idx, scores = ica.find_bads_ecg(raw, method='ctps')
     assert_equal(len(scores), ica.n_components_)
-    idx, scoress = ica.find_bads_eog(raw)
+    idx, scores = ica.find_bads_ecg(raw, method='correlation')
+    assert_equal(len(scores), ica.n_components_)
+    idx, scores = ica.find_bads_ecg(epochs, method='ctps')
+    assert_equal(len(scores), ica.n_components_)
+    assert_raises(ValueError, ica.find_bads_ecg, epochs.average(),
+                  method='ctps')
+    assert_raises(ValueError, ica.find_bads_ecg, raw, method='crazy-coupling')
+
+    idx, scores = ica.find_bads_eog(raw)
     assert_equal(len(scores), ica.n_components_)
     raw.info['chs'][raw.ch_names.index('EOG 061') - 1]['kind'] = 202
     idx, scores = ica.find_bads_eog(raw)